fixed ioctl arg type

+++ include/char.h
@@ -11,7 +11,7 @@ void init_char_devs();
 chardevice_t *set_chardevice(int maj, int (*f)(int, int, char*, size_t), 
 	int (*c)(int, int, long), int (*s)(int, int));
 int char_rw(int rw, dev_t dev, char *buf, size_t len);
-int char_ioctl(dev_t dev, int cmd, int arg);
+int char_ioctl(dev_t dev, int cmd, long arg);
 int set_availablecd(int (*f)(int, int, char*, size_t), int (*c)(int, int, long), 
 	int (*s)(int, int));
 void unregister_char_device(int n);
